server/requirements.txt
deepface==0.0.79
fastapi==0.95.1
fer==22.5.0
moviepy==1.0.3
numpy==1.21.6
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
pandas==1.3.5
pydantic==1.10.2
pymongo==4.3.3
python-dotenv==0.21.1
scikit_learn==1.0.2
uvicorn==0.21.1
